Ufstool is able to extract files.

+++ utilities/ufstool/block.c
@@ -133,7 +133,7 @@ ufs_seek (ufs_t *disk, off_t offset)
     if (lseek (disk->d_fd, offset, 0) < 0) {
         if (verbose)
             printf ("error seeking %lld, block %d\n",
-                offset, (int) (offset / bsize));
+                (long long)offset, (int) (offset / bsize));
         return -1;
     }
     disk->seek = offset;
@@ -147,7 +147,8 @@ ufs_read8 (ufs_t *disk, unsigned char *val)
 
     if (read (disk->d_fd, val, 1) != 1) {
         if (verbose)
-            printf ("error read8, seek %lld block %d\n", disk->seek, (int) (disk->seek / bsize));
+            printf ("error read8, seek %lld block %d\n",
+                (long long)disk->seek, (int) (disk->seek / bsize));
         return -1;
     }
     return 0;
@@ -161,7 +162,8 @@ ufs_read16 (ufs_t *disk, unsigned short *val)
 
     if (read (disk->d_fd, data, 2) != 2) {
         if (verbose)
-            printf ("error read16, seek %lld block %d\n", disk->seek, (int) (disk->seek / bsize));
+            printf ("error read16, seek %lld block %d\n",
+                (long long)disk->seek, (int) (disk->seek / bsize));
         return -1;
     }
     *val = data[1] << 8 | data[0];
@@ -176,7 +178,8 @@ ufs_read32 (ufs_t *disk, unsigned *val)
 
     if (read (disk->d_fd, data, 4) != 4) {
         if (verbose)
-            printf ("error read32, seek %lld block %d\n", disk->seek, (int) (disk->seek / bsize));
+            printf ("error read32, seek %lld block %d\n",
+                (long long)disk->seek, (int) (disk->seek / bsize));
         return -1;
     }
     *val = (unsigned long) data[0] | (unsigned long) data[1] << 8 |
@@ -192,7 +195,8 @@ ufs_read64 (ufs_t *disk, unsigned long long *val)
 
     if (read (disk->d_fd, data, 8) != 8) {
         if (verbose)
-            printf ("error read32, seek %lld block %d\n", disk->seek, (int) (disk->seek / bsize));
+            printf ("error read32, seek %lld block %d\n",
+                (long long)disk->seek, (int) (disk->seek / bsize));
         return -1;
     }
     *val = (unsigned long) data[0] | (unsigned long) data[1] << 8 |
